restructure alimonitor-services module
