+++ app/lib/alimonitor-services/helpers/Cacher.js
@@ -12,6 +12,7 @@
  * granted to it by virtue of its status as an Intergovernmental Organization
  * or submit itself to any jurisdiction.
  */
+
 const path = require('path');
 const fs = require('fs');
 
@@ -58,6 +59,7 @@ class Cacher {
     static serviceCacheDir(synchronizerName) {
         return path.join(
             __dirname,
+            '..',
             '..',
             '..',
             '..',
